+++ sdk/containerservice/arm-containerservicefleet/src/api/fleets/index.ts
@@ -1,0 +1,21 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+export {
+  listCredentials,
+  listBySubscription,
+  listByResourceGroup,
+  $delete,
+  updateAsync,
+  create,
+  get,
+} from "./operations.js";
+export {
+  FleetsListCredentialsOptionalParams,
+  FleetsListBySubscriptionOptionalParams,
+  FleetsListByResourceGroupOptionalParams,
+  FleetsDeleteOptionalParams,
+  FleetsUpdateAsyncOptionalParams,
+  FleetsCreateOptionalParams,
+  FleetsGetOptionalParams,
+} from "./options.js";