+++ sdk/storage/storage-queue/samples/v12/typescript/src/proxyAuth.ts
@@ -8,10 +8,9 @@
 import { StorageSharedKeyCredential, QueueServiceClient } from "@azure/storage-queue";
 
 // Load the .env file if it exists
-import * as dotenv from "dotenv";
-dotenv.config();
+import "dotenv/config";
 
-export async function main() {
+export async function main(): Promise<void> {
   // Enter your storage account name and shared key
   const account = process.env.ACCOUNT_NAME || "";
   const accountKey = process.env.ACCOUNT_KEY || "";