+++ sdk/monitor/monitor-opentelemetry/CHANGELOG.md
@@ -1,6 +1,24 @@
 # Release History
 
-## 1.9.1 ()
+## 1.11.0 (2025-04-16)
+
+### Features Added
+
+- Added support for Process Time Normalized and Exception Rate performance counters.
+
+### Bugs Fixed
+
+- Fixed process time normalized calculation returning NaN.
+
+### Other Changes
+
+- Hide iKey in debug logs.
+
+## 1.10.0 (2025-04-09)
+
+### Features Added
+
+- Add support for performance counters.
 
 ## Other Changes